tests/wpt/web-platform-tests/trusted-types/embed-src.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.sub.js"></script>
<body>
<script>
  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.safe);

    var s = document.createElement('embed');
    s.src = url;
    assert_equals(url + '', s.src);
  }, "src = TrustedScriptURL.unsafelyCreate(URLS.safe)");

  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.javascript);

    var s = document.createElement('embed');
    s.src = url;
    assert_equals(url + '', s.src);
  }, "src = TrustedScriptURL.unsafelyCreate(URLS.javascript)");

  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.external);

    var s = document.createElement('embed');
    s.src = url;
    assert_equals(url + '', s.src);
  }, "src = TrustedScriptURL.unsafelyCreate(URLS.external)");
</script>
